requirements.txt
Flask==3.1.3
google-genai==1.66.0
